benchmark/ct_perftest/requirements.txt
click==8.1.7
nixl==0.1.1
numpy==2.2.5
PyYAML==6.0.2
tabulate==0.9.0
torch==2.6.0
tqdm==4.66.5
